+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 
 - Added `Tab` key to key binding info grid
 
+- Added key binding info about scrolling through document and search result list
+
 - Minor fix for `{line_num}` placeholder handling in `--open-with`
 
     - This should always be usable for text files but previously